Pivot from navigating to the Netobserv plugin to instead navigating to OSSMC's sibling tabs, provided by Netobserv. This is done for Workloads and Services, where netobserv will inject its own tabs.
Note that even if we were to navigate to the netobserv plugin, we shouldn't do it
in the way it was being done, which, I think, establishes a new console session
and is very slow, as opposed to navigating within the existing session.

The new approach leverages the same linking approach we use for our own OSSMC
tab navigation.